+++ src/aui-widget/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1655](https://issues.liferay.com/browse/AUI-1655) WidgetTrigger doesn't work with `tap` event
 * [AUI-1526](https://issues.liferay.com/browse/AUI-1526) Create a menu module
 
 ## [3.0.0pr2](https://github.com/liferay/alloy-ui/releases/tag/3.0.0pr2)